mstakx/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-filter==2.1.0
django-filters==0.2.1
django-nose==1.4.6
djangorestframework==3.9.4
idna==2.8
nose==1.3.7
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3